added the float

+++ week01/Exercise4/Program.cs
@@ -37,7 +37,7 @@ class Program
             sum += numbers[i];
         }
         Console.WriteLine($"The sum is: {sum}");
-        int average = sum / numbers.Count;
+        float average = ((float)sum) / numbers.Count;
         Console.WriteLine($"The average is: {average}");
         Console.WriteLine($"The largest number is: {largest}");
     }
